pyproject.toml
[tool.poetry]
name = "online-platform"
version = "0.1.0"
description = ""
authors = ["nikitt <nikittaosipov@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
flake8 = "^7.1.1"
isort = "^5.13.2"
pillow = "^11.0.0"
python-dotenv = "^1.0.1"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
